downloads-generation/data_references/requirements.txt
shellinford
biopython
gtfparse

